Use correct formula for dbFS, change calibration value

+++ MMDVM-Multi.cpp
@@ -91,7 +91,7 @@ int main(int argc, char** argv)
     unsigned int sample_rate = std::max<unsigned int>(conf.getSampleRate(), 125000U);
     unsigned int channel_spacing = 25000U;
     float baseband_shift = 12500.0f;
-    float power_calibration = float(conf.getRSSICalibration());
+    unsigned int power_calibration = conf.getRSSICalibration();
     if((sample_rate % channel_spacing) != 0)
     {
         ::fprintf(stderr, "MMDVM-Multi: Sample Rate must be a multiple of channel space\n");
